修改代码编译告警

+++ frameworks/core/components_ng/gestures/recognizers/swipe_recognizer.cpp
@@ -135,13 +135,13 @@ void SwipeRecognizer::HandleTouchUpEvent(const TouchEvent& event)
         auto seconds = duration.count();
         resultSpeed_ = offset.GetDistance() / seconds;
         if (resultSpeed_ < speed_) {
-            if (currentFingers_ - 1 + matchedTouch_.size() < fingers_) {
+            if (currentFingers_ - 1 + static_cast<int32_t>(matchedTouch_.size()) < fingers_) {
                 LOGI("the result speed %{public}f is less than duration %{public}f", resultSpeed_, speed_);
                 Adjudicate(AceType::Claim(this), GestureDisposal::REJECT);
             }
         } else {
             matchedTouch_.insert(event.id);
-            if (matchedTouch_.size() == fingers_) {
+            if (static_cast<int32_t>(matchedTouch_.size()) == fingers_) {
                 LOGI("try to accepte swipe gesture");
                 Adjudicate(AceType::Claim(this), GestureDisposal::ACCEPT);
             } else {
